build: extract helper test functions
